backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.1
pydantic==2.6.0
google-api-python-client==2.114.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
httpx==0.26.0
python-multipart==0.0.7 